Fixed landing zero for joystick

+++ Processing/Survival_of_the_dumbest/Player.pde
@@ -35,7 +35,7 @@ class Player {
     popMatrix();
   }
 
-  void move(int _x, int _y) {
+  void move(float _x, float _y) {
     redraw();
     velocityX+=accelerationX;
     velocityY+=accelerationY;
